requirements.txt
fastapi==0.95.2
pydantic==1.9.1
stripe==8.6.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
typing-extensions==4.7.1
